Add logger checking for existance of setters to classes
Add a new logger call if a setter method for a property that should have
one is not found for any of the OXP core classes.

Also changes the log level from warning to info for properties of an unkown
trigger.

Fixed formatting in VirtualLocations.php

Removed an unnecessary check from Alert.php that could have lead to silent
discarding of Trigger properties.